+++ src/modules/run_build.js
@@ -1,7 +1,9 @@
 const { Spinner } = require('cli-spinner')
 const { KareeProjectConfig} = require('../models/karee_config')
 const cmd = require('node-cmd')
+const path = require('path')
 const { exit } = require('yargs')
+const io = require('../io/io')
 
 
 class KareeSourceGen{
@@ -15,6 +17,9 @@ class KareeSourceGen{
         spinner.setSpinnerTitle(`\x1b[32m\x1b[1mKaree is generating additional source in \x1b[33m${this.projectConfig.appName}\x1b[39m\x1b[22m\x1b[0m`)
         spinner.start()
 
+        
+        io.delete(`lib${path.sep}core${path.sep}extensions`)
+        io.delete(`lib${path.sep}core${path.sep}core.reflectable.dart`)
         cmd.run('flutter clean && flutter pub get && flutter packages pub run build_runner watch --delete-conflicting-outputs')
            .on('close', (code, signal) => {
                spinner.stop(false)